pyproject.toml
[project]
name = "weather-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Никита Быстров",email = "hallnik96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
pytest = "^8.4.0"
pytest-django = "^4.11.1"
ruff = "^0.11.13"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "preserve"
